fix(smartcartlink): fix bug of always returning either null or cart items count
Fix the bug that makes the SmartCartLink component to only ever display nothing (return null) or the
number of items in cart. This was caused by the order of the different notifications that should be
displayed according to different states in retrieving the number of items in the cart. Re-ordering
the notifications fixed the issue.