fix(cart reducer): fix console error when we click on the 'remove product from cart' button
Fix console error when we click on the 'remove product from cart' button. The error was caused
because we were concatenating the product object, rather than an array of the product object, to the
'removeProductFromCartList' array when the REMOVE_PRODUCT_FROM_CART action is dispatched. So, when
the REMOVE_PRODUCT_FROM_CART_SUCCESS action is dispatched, and we attempt to remove the product from
the array, it is not there.