fix(logo): fix logo not showing on some routes
Fix logo not displaying on routes with extra forward slashes (e.g categories/:category, cart/). The
bug was caused because we used a relative url for referencing our logo, so it tries to load the logo
image from the current path, rather than from the root of the application. Changing our logo
component to reference the image using an absolute url fixed it.